include get_n_columns in the utils
